+++ Lib/fontbakery/profiles/universal.py
@@ -880,29 +880,39 @@ def com_google_fonts_check_valid_glyphnames(ttFont, config):
 def com_google_fonts_check_unique_glyphnames(ttFont):
     """Font contains unique glyph names?"""
     if (
-        ttFont.sfntVersion == b"\x00\x01\x00\x00"
+        ttFont.sfntVersion == "\x00\x01\x00\x00"
         and ttFont.get("post")
-        and ttFont["post"].formatType == 3.0
+        and ttFont["post"].formatType == 3
     ):
         yield SKIP, (
-            "TrueType fonts with a format 3.0 post table contain no glyph names."
+            "TrueType fonts with a format 3 post table contain no glyph names."
+        )
+    elif (
+        ttFont.sfntVersion == "OTTO"
+        and ttFont.get("CFF2")
+        and ttFont.get("post")
+        and ttFont["post"].formatType == 3
+    ):
+        yield SKIP, (
+            "OpenType-CFF2 fonts with a format 3 post table contain no glyph names."
         )
     else:
-        glyphs = []
-        duplicated_glyphIDs = []
-        for _, g in enumerate(ttFont.getGlyphOrder()):
-            glyphID = re.sub(r"#\w+", "", g)
-            if glyphID in glyphs:
-                duplicated_glyphIDs.append(glyphID)
+        glyph_names = set()
+        dup_glyph_names = set()
+        for gname in ttFont.getGlyphOrder():
+            # On font load, Fonttools adds #1, #2, ... suffixes to duplicate glyph names
+            glyph_name = re.sub(r"#\w+", "", gname)
+            if glyph_name in glyph_names:
+                dup_glyph_names.add(glyph_name)
             else:
-                glyphs.append(glyphID)
+                glyph_names.add(glyph_name)
 
-        if len(duplicated_glyphIDs) == 0:
-            yield PASS, "Font contains unique glyph names."
+        if not dup_glyph_names:
+            yield PASS, "Glyph names are all unique."
         else:
             yield FAIL, Message(
                 "duplicated-glyph-names",
-                f"The following glyph names occur twice: {duplicated_glyphIDs}",
+                f"These glyph names occur more than once: {sorted(dup_glyph_names)}",
             )